+++ frontend/src/components/Achievements/Achievements.jsx
@@ -6,8 +6,7 @@ import '../Home/Body/Calender/styles/CalenderWeekdays.css';
 import '../Home/Body/Calender/styles/CalenderNavigation.css';
 import '../Home/Body/Calender/styles/CalenderDays.css';
 import '../Home/Body/Calender/styles/CalenderCommon.css';
-import CalenderFormatShortWeekday from "../Home/Body/Calender/CalenderFormatShortWeekday";
-import CalenderTileContent from "../Home/Body/Calender/CalenderTileContent";
+import { CalenderFormatShortWeekday, CalenderTileContent } from "../Home/Body/Calender/Calender";
 import './styles/Achievements.css';
 import DailyStats from './DailyStats';
